@@ -209,87 +212,89 @@
     "testRegex": "tests/.*\\.test\\.js$"
   },
   "dependencies": {
-    "babel-polyfill": "6.20.0",
+    "babel-polyfill": "6.23.0",
     "chalk": "1.1.3",
     "compression": "1.6.2",
-    "cross-env": "3.1.3",
-    "express": "4.14.0",
-    "fontfaceobserver": "2.0.7",
+    "cross-env": "5.0.0",
+    "express": "4.15.3",
+    "fontfaceobserver": "2.0.9",
+    "history": "4.6.3",
+    "hoist-non-react-statics": "2.1.1",
     "immutable": "3.8.1",
     "intl": "1.2.5",
     "invariant": "2.2.2",
-    "ip": "1.1.4",
-    "lodash": "4.17.2",
+    "ip": "1.1.5",
+    "lodash": "4.17.4",
     "minimist": "1.2.0",
-    "react": "15.4.1",
-    "react-dom": "15.4.1",
-    "react-helmet": "3.2.2",
-    "react-intl": "2.1.5",
-    "react-redux": "4.4.6",
-    "react-router": "3.0.0",
-    "react-router-redux": "4.0.6",
-    "react-router-scroll": "0.4.1",
+    "prop-types": "15.5.10",
+    "react": "15.6.1",
+    "react-dom": "15.6.1",
+    "react-helmet": "5.1.3",
+    "react-intl": "2.3.0",
+    "react-loadable": "4.0.3",
+    "react-redux": "5.0.5",
+    "react-router-dom": "4.1.1",
+    "react-router-redux": "5.0.0-alpha.6",
     "redux": "3.6.0",
-    "redux-immutable": "3.0.8",
-    "redux-saga": "0.14.0",
-    "reselect": "2.5.4",
+    "redux-immutable": "4.0.0",
+    "redux-saga": "0.15.3",
+    "reselect": "3.0.1",
     "sanitize.css": "4.1.0",
-    "styled-components": "1.1.2",
+    "styled-components": "2.0.0",
     "warning": "3.0.0",
-    "whatwg-fetch": "2.0.1"
+    "whatwg-fetch": "2.0.3"
   },
   "devDependencies": {
-    "babel-cli": "6.18.0",
-    "babel-core": "6.21.0",
-    "babel-eslint": "7.1.1",
-    "babel-loader": "6.2.10",
-    "babel-plugin-dynamic-import-node": "1.0.0",
-    "babel-plugin-react-intl": "2.2.0",
+    "add-asset-html-webpack-plugin": "2.0.1",
+    "babel-cli": "6.24.1",
+    "babel-core": "6.24.1",
+    "babel-eslint": "7.2.3",
+    "babel-loader": "7.1.0",
+    "babel-plugin-dynamic-import-node": "1.0.2",
+    "babel-plugin-react-intl": "2.3.1",
     "babel-plugin-react-transform": "2.0.2",
-    "babel-plugin-transform-es2015-modules-commonjs": "6.18.0",
-    "babel-plugin-transform-react-constant-elements": "6.9.1",
-    "babel-plugin-transform-react-inline-elements": "6.8.0",
-    "babel-plugin-transform-react-remove-prop-types": "0.2.11",
-    "babel-preset-latest": "6.16.0",
-    "babel-preset-react": "6.16.0",
-    "babel-preset-react-hmre": "1.1.1",
-    "babel-preset-stage-0": "6.16.0",
-    "cheerio": "0.22.0",
-    "circular-dependency-plugin": "2.0.0",
-    "coveralls": "2.11.15",
-    "css-loader": "0.26.1",
-    "enzyme": "2.6.0",
-    "eslint": "3.11.1",
-    "eslint-config-airbnb": "13.0.0",
-    "eslint-config-airbnb-base": "10.0.1",
-    "eslint-import-resolver-webpack": "0.8.0",
-    "eslint-plugin-import": "2.2.0",
-    "eslint-plugin-jsx-a11y": "2.2.3",
-    "eslint-plugin-react": "6.7.1",
-    "eslint-plugin-redux-saga": "0.1.5",
+    "babel-plugin-styled-components": "1.1.4",
+    "babel-plugin-transform-es2015-modules-commonjs": "6.24.1",
+    "babel-plugin-transform-react-constant-elements": "6.23.0",
+    "babel-plugin-transform-react-inline-elements": "6.22.0",
+    "babel-plugin-transform-react-remove-prop-types": "0.4.5",
+    "babel-preset-env": "1.5.1",
+    "babel-preset-react": "6.24.1",
+    "babel-preset-stage-0": "6.24.1",
+    "circular-dependency-plugin": "3.0.0",
+    "coveralls": "2.13.1",
+    "css-loader": "0.28.4",
+    "enzyme": "2.8.2",
+    "eslint": "3.19.0",
+    "eslint-config-airbnb": "15.0.1",
+    "eslint-config-airbnb-base": "11.2.0",
+    "eslint-import-resolver-webpack": "0.8.3",
+    "eslint-plugin-import": "2.7.0",
+    "eslint-plugin-jsx-a11y": "5.0.3",
+    "eslint-plugin-react": "7.0.1",
+    "eslint-plugin-redux-saga": "0.3.0",
     "eventsource-polyfill": "0.9.6",
-    "exports-loader": "0.6.3",
-    "file-loader": "0.9.0",
-    "html-loader": "0.4.4",
-    "html-webpack-plugin": "2.24.1",
+    "exports-loader": "0.6.4",
+    "file-loader": "0.11.1",
+    "html-loader": "0.4.5",
+    "html-webpack-plugin": "2.29.0",
     "image-webpack-loader": "2.0.0",
-    "imports-loader": "0.6.5",
-    "jest-cli": "18.0.0",
-    "lint-staged": "3.2.1",
-    "ngrok": "2.2.4",
-    "node-plop": "0.5.4",
+    "imports-loader": "0.7.1",
+    "jest-cli": "20.0.4",
+    "lint-staged": "3.5.1",
+    "ngrok": "2.2.9",
+    "node-plop": "0.7.0",
     "null-loader": "0.1.1",
-    "offline-plugin": "4.5.2",
-    "plop": "1.7.3",
-    "pre-commit": "1.1.3",
-    "react-addons-test-utils": "15.4.1",
-    "rimraf": "2.5.4",
-    "shelljs": "0.7.5",
-    "sinon": "2.0.0-pre",
-    "style-loader": "0.13.1",
-    "url-loader": "0.5.7",
-    "webpack": "2.2.0-rc.3",
-    "webpack-dev-middleware": "1.9.0",
-    "webpack-hot-middleware": "2.15.0"
+    "offline-plugin": "4.8.1",
+    "plop": "1.8.0",
+    "pre-commit": "1.2.2",
+    "react-test-renderer": "15.6.1",
+    "rimraf": "2.6.1",
+    "shelljs": "0.7.7",
+    "style-loader": "0.18.1",
+    "url-loader": "0.5.8",
+    "webpack": "3.5.5",
+    "webpack-dev-middleware": "1.11.0",
+    "webpack-hot-middleware": "2.18.0"
   }
 }
