@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 import Ul from './Ul';
 import Wrapper from './Wrapper';
@@ -9,8 +10,8 @@ function List(props) {
 
   // If we have items, render them
   if (props.items) {
-    content = props.items.map((item, index) => (
-      <ComponentToRender key={`item-${index}`} item={item} />
+    content = props.items.map((item) => (
+      <ComponentToRender key={`item-${item.id}`} item={item} />
     ));
   } else {
     // Otherwise render a single component
@@ -27,8 +28,8 @@ function List(props) {
 }
 
 List.propTypes = {
-  component: React.PropTypes.func.isRequired,
-  items: React.PropTypes.array,
+  component: PropTypes.func.isRequired,
+  items: PropTypes.array,
 };
 
 export default List;
