@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 import Item from './Item';
 import Wrapper from './Wrapper';
@@ -14,7 +15,7 @@ function ListItem(props) {
 }
 
 ListItem.propTypes = {
-  item: React.PropTypes.any,
+  item: PropTypes.any,
 };
 
 export default ListItem;
