@@ -5,6 +5,7 @@
  */
 
 import React from 'react';
+import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { createSelector } from 'reselect';
 
@@ -26,8 +27,8 @@ export class LocaleToggle extends React.PureComponent { // eslint-disable-line r
 }
 
 LocaleToggle.propTypes = {
-  onLocaleToggle: React.PropTypes.func,
-  locale: React.PropTypes.string,
+  onLocaleToggle: PropTypes.func,
+  locale: PropTypes.string,
 };
 
 const mapStateToProps = createSelector(
