@@ -1,12 +1,13 @@
 /**
  *
- * Button.react.js
+ * Button.js
  *
  * A common button, if you pass it a prop "route" it'll render a link to a react-router route
  * otherwise it'll render a link with an onclick
  */
 
-import React, { PropTypes, Children } from 'react';
+import React, { Children } from 'react';
+import PropTypes from 'prop-types';
 
 import A from './A';
 import StyledButton from './StyledButton';
