@@ -1,11 +1,12 @@
 /**
  *
- * Img.react.js
+ * Img.js
  *
  * Renders an image, enforcing the usage of the alt="" tag
  */
 
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 
 function Img(props) {
   return (
