@@ -1,25 +1,30 @@
 import React from 'react';
-import { render } from 'enzyme';
+import { shallow } from 'enzyme';
 
 import ListItem from 'components/ListItem';
 import List from '../index';
 
 describe('<List />', () => {
   it('should render the component if no items are passed', () => {
-    const renderedComponent = render(
+    const renderedComponent = shallow(
       <List component={ListItem} />
     );
     expect(renderedComponent.find(ListItem)).toBeDefined();
   });
 
-  it('should render the items', () => {
+  it('should pass all items props to rendered component', () => {
     const items = [
-      'Hello',
-      'World',
+      { id: 1, name: 'Hello' },
+      { id: 2, name: 'World' },
     ];
-    const renderedComponent = render(
-      <List items={items} component={ListItem} />
+
+    const component = ({ item }) => <ListItem>{item.name}</ListItem>; // eslint-disable-line react/prop-types
+
+    const renderedComponent = shallow(
+      <List items={items} component={component} />
     );
-    expect(renderedComponent.find(items)).toBeDefined();
+    expect(renderedComponent.find(component)).toHaveLength(2);
+    expect(renderedComponent.find(component).at(0).prop('item')).toBe(items[0]);
+    expect(renderedComponent.find(component).at(1).prop('item')).toBe(items[1]);
   });
 });
