@@ -1,9 +1,10 @@
 import React from 'react';
 import { shallow } from 'enzyme';
+import { Route } from 'react-router-dom';
 
 import Header from 'components/Header';
 import Footer from 'components/Footer';
-import { App } from '../index';
+import App from '../index';
 
 describe('<App />', () => {
   it('should render the header', () => {
@@ -13,14 +14,11 @@ describe('<App />', () => {
     expect(renderedComponent.find(Header).length).toBe(1);
   });
 
-  it('should render its children', () => {
-    const children = (<h1>Test</h1>);
+  it('should render some routes', () => {
     const renderedComponent = shallow(
-      <App>
-        {children}
-      </App>
+      <App />
     );
-    expect(renderedComponent.contains(children)).toBe(true);
+    expect(renderedComponent.find(Route).length).not.toBe(0);
   });
 
   it('should render the footer', () => {
