@@ -4,7 +4,7 @@
  * List all the features
  */
 import React from 'react';
-import Helmet from 'react-helmet';
+import { Helmet } from 'react-helmet';
 import { FormattedMessage } from 'react-intl';
 
 import H1 from 'components/H1';
@@ -24,12 +24,10 @@ export default class FeaturePage extends React.Component { // eslint-disable-lin
   render() {
     return (
       <div>
-        <Helmet
-          title="Feature Page"
-          meta={[
-            { name: 'description', content: 'Feature page of React.js Boilerplate application' },
-          ]}
-        />
+        <Helmet>
+          <title>Feature Page</title>
+          <meta name="description" content="Feature page of React.js Boilerplate application" />
+        </Helmet>
         <H1>
           <FormattedMessage {...messages.header} />
         </H1>
