@@ -33,6 +33,33 @@ describe('<HomePage />', () => {
     expect(submitSpy).toHaveBeenCalled();
   });
 
+  it('should not call onSubmitForm if username is an empty string', () => {
+    const submitSpy = jest.fn();
+    mount(
+      <IntlProvider locale="en">
+        <HomePage
+          onChangeUsername={() => {}}
+          onSubmitForm={submitSpy}
+        />
+      </IntlProvider>
+    );
+    expect(submitSpy).not.toHaveBeenCalled();
+  });
+
+  it('should not call onSubmitForm if username is null', () => {
+    const submitSpy = jest.fn();
+    mount(
+      <IntlProvider locale="en">
+        <HomePage
+          username=""
+          onChangeUsername={() => {}}
+          onSubmitForm={submitSpy}
+        />
+      </IntlProvider>
+    );
+    expect(submitSpy).not.toHaveBeenCalled();
+  });
+
   describe('mapDispatchToProps', () => {
     describe('onChangeUsername', () => {
       it('should be injected', () => {
@@ -49,28 +76,28 @@ describe('<HomePage />', () => {
         expect(dispatch).toHaveBeenCalledWith(changeUsername(username));
       });
     });
-  });
 
-  describe('onSubmitForm', () => {
-    it('should be injected', () => {
-      const dispatch = jest.fn();
-      const result = mapDispatchToProps(dispatch);
-      expect(result.onSubmitForm).toBeDefined();
-    });
+    describe('onSubmitForm', () => {
+      it('should be injected', () => {
+        const dispatch = jest.fn();
+        const result = mapDispatchToProps(dispatch);
+        expect(result.onSubmitForm).toBeDefined();
+      });
 
-    it('should dispatch loadRepos when called', () => {
-      const dispatch = jest.fn();
-      const result = mapDispatchToProps(dispatch);
-      result.onSubmitForm();
-      expect(dispatch).toHaveBeenCalledWith(loadRepos());
-    });
+      it('should dispatch loadRepos when called', () => {
+        const dispatch = jest.fn();
+        const result = mapDispatchToProps(dispatch);
+        result.onSubmitForm();
+        expect(dispatch).toHaveBeenCalledWith(loadRepos());
+      });
 
-    it('should preventDefault if called with event', () => {
-      const preventDefault = jest.fn();
-      const result = mapDispatchToProps(() => {});
-      const evt = { preventDefault };
-      result.onSubmitForm(evt);
-      expect(preventDefault).toHaveBeenCalledWith();
+      it('should preventDefault if called with event', () => {
+        const preventDefault = jest.fn();
+        const result = mapDispatchToProps(() => {});
+        const evt = { preventDefault };
+        result.onSubmitForm(evt);
+        expect(preventDefault).toHaveBeenCalledWith();
+      });
     });
   });
 });
