@@ -1,8 +1,6 @@
-import styled from 'styled-components';
+import NormalA from 'components/A';
 
-import A from 'components/A';
-
-const IssueLink = styled(A)`
+const IssueLink = NormalA.extend`
   height: 100%;
   color: black;
   display: flex;
