@@ -1,58 +1,57 @@
-require('shelljs/global');
+const shell = require('shelljs');
 const addCheckMark = require('./helpers/checkmark.js');
 
-if (!which('git')) {
-  echo('Sorry, this script requires git');
-  exit(1);
+if (!shell.which('git')) {
+  shell.echo('Sorry, this script requires git');
+  shell.exit(1);
 }
 
-if (!test('-e', 'internals/templates')) {
-  echo('The example is deleted already.');
-  exit(1);
+if (!shell.test('-e', 'internals/templates')) {
+  shell.echo('The example is deleted already.');
+  shell.exit(1);
 }
 
 process.stdout.write('Cleanup started...');
 
 // Reuse existing LanguageProvider and i18n tests
-mv('app/containers/LanguageProvider/tests', 'internals/templates/containers/LanguageProvider');
-cp('app/tests/i18n.test.js', 'internals/templates/tests/i18n.test.js');
+shell.mv('app/containers/LanguageProvider/tests', 'internals/templates/containers/LanguageProvider');
+shell.cp('app/tests/i18n.test.js', 'internals/templates/tests/i18n.test.js');
 
 // Cleanup components/
-rm('-rf', 'app/components/*');
+shell.rm('-rf', 'app/components/*');
 
 // Handle containers/
-rm('-rf', 'app/containers');
-mv('internals/templates/containers', 'app');
+shell.rm('-rf', 'app/containers');
+shell.mv('internals/templates/containers', 'app');
 
 // Handle tests/
-mv('internals/templates/tests', 'app');
+shell.mv('internals/templates/tests', 'app');
 
 // Handle translations/
-rm('-rf', 'app/translations')
-mv('internals/templates/translations', 'app');
+shell.rm('-rf', 'app/translations')
+shell.mv('internals/templates/translations', 'app');
 
 // Handle utils/
-rm('-rf', 'app/utils');
-mv('internals/templates/utils', 'app')
+shell.rm('-rf', 'app/utils');
+shell.mv('internals/templates/utils', 'app')
 
 // Replace the files in the root app/ folder
-cp('internals/templates/app.js', 'app/app.js');
-cp('internals/templates/global-styles.js', 'app/global-styles.js');
-cp('internals/templates/i18n.js', 'app/i18n.js');
-cp('internals/templates/index.html', 'app/index.html');
-cp('internals/templates/reducers.js', 'app/reducers.js');
-cp('internals/templates/routes.js', 'app/routes.js');
-cp('internals/templates/store.js', 'app/store.js');
+shell.cp('internals/templates/app.js', 'app/app.js');
+shell.cp('internals/templates/global-styles.js', 'app/global-styles.js');
+shell.cp('internals/templates/i18n.js', 'app/i18n.js');
+shell.cp('internals/templates/index.html', 'app/index.html');
+shell.cp('internals/templates/reducers.js', 'app/reducers.js');
+shell.cp('internals/templates/configureStore.js', 'app/configureStore.js');
 
 // Remove the templates folder
-rm('-rf', 'internals/templates');
+shell.rm('-rf', 'internals/templates');
 
 addCheckMark();
 
 // Commit the changes
-if (exec('git add . --all && git commit -qm "Remove default example"').code !== 0) {
-  echo('\nError: Git commit failed');
-  exit(1);
+if (shell.exec('git add . --all && git commit -qm "Remove default example"').code !== 0) {
+  shell.echo('\nError: Git commit failed');
+  shell.exit(1);
 }
 
-echo('\nCleanup done. Happy Coding!!!');
+shell.echo('\nCleanup done. Happy Coding!!!');
