@@ -14,11 +14,10 @@ process.stdin.resume();
 process.stdin.setEncoding('utf8');
 
 process.stdout.write('\n');
-let interval = animateProgress('Cleaning old repository');
-process.stdout.write('Cleaning old repository');
+let interval;
+let clearRepo = true;
 
-cleanRepo(function () {
-  clearInterval(interval);
+cleanRepo(() => {
   process.stdout.write('\nInstalling dependencies... (This might take a while)');
   setTimeout(function () {
     readline.cursorTo(process.stdout, 0);
@@ -29,10 +28,42 @@ cleanRepo(function () {
 });
 
 /**
- * Deletes the .git folder in dir
+ * Deletes the .git folder in dir only if cloned from our repo
  */
 function cleanRepo(callback) {
-  shell.rm('-rf', '.git/');
+  fs.readFile('.git/config', 'utf8', (err, data) => {
+    if(!err) {
+      let isClonedRepo = typeof data === 'string' 
+        && (data.match(/url\s*=/g) || []).length === 1
+        && /react-boilerplate\/react-boilerplate\.git/.test(data);
+      if(isClonedRepo) {
+        process.stdout.write('\nDo you want to clear old repository? [Y/n] ');
+        process.stdin.resume();
+        process.stdin.on('data', (data) => {
+          let val = data.toString().trim();
+          if(val === 'y' || val === 'Y' || val === '') {
+            process.stdout.write('Removing old repository');
+            shell.rm('-rf', '.git/');
+            addCheckMark(callback);
+          } else {
+            dontClearRepo('', callback); 
+          }
+        });
+      } else {
+        dontClearRepo('\n', callback);
+      }
+    } else {
+      callback();
+    }
+  });
+}
+
+/**
+ * Function which indicates that we are not cleaning git repo
+ */
+function dontClearRepo(nl, callback) {
+  clearRepo = false;
+  process.stdout.write(nl + 'Leaving your repository untouched');
   addCheckMark(callback);
 }
 
@@ -83,12 +114,14 @@ function installDepsCallback(error) {
   }
 
   deleteFileInCurrentDir('setup.js', function () {
-    interval = animateProgress('Initialising new repository');
-    process.stdout.write('Initialising new repository');
-    initGit(function () {
-      clearInterval(interval);
-      process.stdout.write('\nDone!');
-      process.exit(0);
-    });
+    if(clearRepo) {
+      interval = animateProgress('Initialising new repository');
+      process.stdout.write('Initialising new repository');
+      initGit(function () {
+        clearInterval(interval);
+      });
+    }
+    process.stdout.write('\nDone!');
+    process.exit(0);
   });
 }
