@@ -1,20 +1,12 @@
-// makeSelectLocationState expects a plain JS object for the routing state
-const makeSelectLocationState = () => {
-  let prevRoutingState;
-  let prevRoutingStateJS;
+import { createSelector } from 'reselect';
 
-  return (state) => {
-    const routingState = state.get('route'); // or state.route
+const selectRoute = (state) => state.get('route');
 
-    if (!routingState.equals(prevRoutingState)) {
-      prevRoutingState = routingState;
-      prevRoutingStateJS = routingState.toJS();
-    }
-
-    return prevRoutingStateJS;
-  };
-};
+const makeSelectLocation = () => createSelector(
+  selectRoute,
+  (routeState) => routeState.get('location').toJS()
+);
 
 export {
-  makeSelectLocationState,
+  makeSelectLocation,
 };
