@@ -1,16 +1,14 @@
 import React from 'react';
 import { shallow } from 'enzyme';
+import { Route } from 'react-router-dom';
 
 import App from '../index';
 
 describe('<App />', () => {
-  it('should render its children', () => {
-    const children = (<h1>Test</h1>);
+  it('should render some routes', () => {
     const renderedComponent = shallow(
-      <App>
-        {children}
-      </App>
+      <App />
     );
-    expect(renderedComponent.contains(children)).toBe(true);
+    expect(renderedComponent.find(Route).length).not.toBe(0);
   });
 });
