@@ -4,26 +4,38 @@
 
 const path = require('path');
 const fs = require('fs');
+const glob = require('glob');
 const webpack = require('webpack');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
+const AddAssetHtmlPlugin = require('add-asset-html-webpack-plugin');
 const CircularDependencyPlugin = require('circular-dependency-plugin');
 const logger = require('../../server/logger');
-const cheerio = require('cheerio');
 const pkg = require(path.resolve(process.cwd(), 'package.json'));
 const dllPlugin = pkg.dllPlugin;
 
 const plugins = [
   new webpack.HotModuleReplacementPlugin(), // Tell webpack we want hot reloading
-  new webpack.NoErrorsPlugin(),
+  new webpack.NoEmitOnErrorsPlugin(),
   new HtmlWebpackPlugin({
     inject: true, // Inject all files that are generated by webpack, e.g. bundle.js
-    templateContent: templateContent(), // eslint-disable-line no-use-before-define
+    template: 'app/index.html',
   }),
   new CircularDependencyPlugin({
     exclude: /a\.js|node_modules/, // exclude node_modules
     failOnError: false, // show a warning when there is a circular dependency
   }),
 ];
+
+if (dllPlugin) {
+  glob.sync(`${dllPlugin.path}/*.dll.js`).forEach((dllPath) => {
+    plugins.push(
+      new AddAssetHtmlPlugin({
+        filepath: dllPath,
+        includeSourcemap: false,
+      })
+    );
+  });
+}
 
 module.exports = require('./webpack.base.babel')({
   // Add hot reloading in development
@@ -42,17 +54,9 @@ module.exports = require('./webpack.base.babel')({
   // Add development plugins
   plugins: dependencyHandlers().concat(plugins), // eslint-disable-line no-use-before-define
 
-  // Tell babel that we want to hot-reload
-  babelQuery: {
-    // require.resolve solves the issue of relative presets when dealing with
-    // locally linked packages. This is an issue with babel and webpack.
-    // See https://github.com/babel/babel-loader/issues/149 and
-    // https://github.com/webpack/webpack/issues/1866
-    presets: ['babel-preset-react-hmre'].map(require.resolve),
-  },
-
   // Emit a source map for easier debugging
-  devtool: 'cheap-module-eval-source-map',
+  // See https://webpack.js.org/configuration/devtool/#devtool
+  devtool: 'eval-source-map',
 
   performance: {
     hints: false,
@@ -125,24 +129,4 @@ function dependencyHandlers() {
       manifest: require(manifestPath), // eslint-disable-line global-require
     });
   });
-}
-
-/**
- * We dynamically generate the HTML content in development so that the different
- * DLL Javascript files are loaded in script tags and available to our application.
- */
-function templateContent() {
-  const html = fs.readFileSync(
-    path.resolve(process.cwd(), 'app/index.html')
-  ).toString();
-
-  if (!dllPlugin) { return html; }
-
-  const doc = cheerio(html);
-  const body = doc.find('body');
-  const dllNames = !dllPlugin.dlls ? ['reactBoilerplateDeps'] : Object.keys(dllPlugin.dlls);
-
-  dllNames.forEach((dllName) => body.append(`<script data-dll='true' src='/${dllName}.dll.js'></script>`));
-
-  return doc.toString();
 }
